+++ skypro-music/src/components/SignUp/SignUp.module.css
@@ -1,0 +1,153 @@
+.block {
+    position: absolute;
+    z-index: 2;
+    left: calc(50% - (366px / 2));
+    top: calc(50% - (439px / 2));
+    opacity: 1;
+}
+
+.formLogin {
+    width: 366px;
+    min-height: 489px;
+    background-color: #ffffff;
+    border-radius: 12px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
+    flex-direction: column;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    padding: 43px 47px 47px 40px;
+}
+
+.logo {
+    width: 140px;
+    height: 21px;
+    margin-bottom: 34px;
+    background-color: transparent;
+
+}
+
+.logoImage {
+    width: 140px;
+    height: auto;
+}
+
+.input, .inputFirstChild {
+    width: 100%;
+    border-top: none;
+    border-left: none;
+    border-right: none;
+    border-bottom: 1px solid #d0cece;
+    padding: 8px 10px;
+    margin-right: 3px;
+    margin-bottom: 30px;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+
+    &::placeholder {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 24px;
+        letter-spacing: -0.05px;
+        color: #d0cece;
+    }
+
+    &::-webkit-input-placeholder {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 24px;
+        letter-spacing: -0.05px;
+        color: #d0cece;
+    }
+
+    &:-ms-input-placeholder {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 24px;
+        letter-spacing: -0.05px;
+        color: #d0cece;
+    }
+
+    &::-ms-input-placeholder {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 18px;
+        line-height: 24px;
+        letter-spacing: -0.05px;
+        color: #d0cece;
+    }
+}
+
+.inputFirstChild {
+    margin-bottom: 30px;
+}
+
+.signupButton {
+    width: 278px;
+    height: 62px;
+    background-color: #580EA2;
+    border-radius: 6px;
+    margin-left: 4px;
+    border: none;
+    margin-top: 30px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+
+    &:hover {
+        background-color: #3f007d;
+    }
+}
+
+.signupButtonActive {
+    background-color: #271A58;
+}
+
+.signupButtonLink {
+    width: 100%;
+    height: 100%;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #FFFFFF;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+}
+
+.errorBox {
+    height: 30px;
+    margin-left: 5px;
+    margin-right: 5px;
+    text-align: center;
+}
+
+.error {
+    color: red;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+}
